plat/msdos386: fix some issues in libsys code
 - ensure es = ds = ss upon exit of each C runtime function
 - clear upper 16 bits of ebx before setting them to 0x0021,
   when invoking interrupt_ptr to simulate a RM int 0x21
 - make _sys_exists use the transfer buffer (which it needs)
 - make _sys_rawread properly handle an end-of-file read
   (zero bytes read)
 - make argument to _sys_seterrno short --- after a failed
   int 0x21 call, only the lower 16 bits of eax hold the MS-DOs
   error code
 - _sys_rawlseek accepts only 3 longword arguments, not 4
   (the offset is only 1 longword)
 - other minor fixes

+++ plat/msdos/libsys/sys_seterrno.c
@@ -35,7 +35,7 @@ static const signed char err_map[] =
  * Map an MS-DOS 2+ system error code to an `errno' value and store that in
  * `errno'.  Return a longword -1.
  */
-long _sys_seterrno(unsigned dos_err)
+long _sys_seterrno(unsigned short dos_err)
 {
 	if (dos_err < sizeof(err_map) / sizeof(err_map[0]))
 		errno = err_map[dos_err];
